+++ src/main/java/com/outstagram/outstagram/config/database/DataSourceRouting.java
@@ -7,7 +7,7 @@ import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
 public class DataSourceRouting extends AbstractRoutingDataSource {
     @Override
     protected Object determineCurrentLookupKey() {
-        Integer shardId = DataSourceContextHolder.getShardId();
+        Long shardId = DataSourceContextHolder.getShardId();
         log.info("=============================== Current Shard ID: {}", shardId);
         return shardId;
     }
